Separting handlers and authenticator files
